Added implementation for complete submission endpoint
